removed floating point battery reading, going with ADC reading. Added maxbotix function, adapted Sysclock changes

+++ FloodnetDriver/Adc.h
@@ -10,7 +10,7 @@
 #include "main.h"
 
 
-float AdcRead_VBatt(void);
+uint16_t AdcRead_VBatt(void);
